requirements.txt
# Runtime dependencies for lanpaint_release
# Tested with Python 3.12 and CUDA-enabled PyTorch in this workspace.

LanPaint==1.4.13
torch==2.8.0
diffusers==0.37.0.dev0
transformers==4.57.6
accelerate==1.12.0
safetensors==0.7.0
numpy==2.3.2
Pillow==11.3.0
tqdm==4.67.1
requests==2.32.5
huggingface-hub==0.36.0
scipy==1.17.0
sentencepiece==0.2.1
protobuf==6.33.4
